Move VectorData and VectorGeometry to a new place.
