+++ Wrappers/Python/cil/framework/framework.py
@@ -2567,146 +2567,6 @@ class ImageData(DataContainer):
             return image_data_out
 
 
-
-class VectorData(DataContainer):
-    '''DataContainer to contain 1D array'''
-
-    @property
-    def geometry(self):
-        return self._geometry
-
-    @geometry.setter
-    def geometry(self, val):
-        self._geometry = val
-
-    @property
-    def dimension_labels(self):
-        if hasattr(self,'geometry'):
-            return self.geometry.dimension_labels
-        else:
-            return self._dimension_labels
-
-    @dimension_labels.setter
-    def dimension_labels(self, val):
-        if hasattr(self,'geometry'):
-            self.geometry.dimension_labels = val
-        
-        self._dimension_labels = val
-
-    def __init__(self, array=None, **kwargs):
-        self.geometry = kwargs.get('geometry', None)
-
-        dtype = kwargs.get('dtype', numpy.float32)
-        
-        if self.geometry is None:
-            if array is None:
-                raise ValueError('Please specify either a geometry or an array')
-            else:
-                if len(array.shape) > 1:
-                    raise ValueError('Incompatible size: expected 1D got {}'.format(array.shape))
-                out = array
-                self.geometry = VectorGeometry(array.shape[0], **kwargs)
-                self.length = self.geometry.length
-        else:
-            self.length = self.geometry.length
-                
-            if array is None:
-                out = numpy.zeros((self.length,), dtype=dtype)
-            else:
-                if self.length == array.shape[0]:
-                    out = array
-                else:
-                    raise ValueError('Incompatible size: expecting {} got {}'.format((self.length,), array.shape))
-        deep_copy = True
-        # need to pass the geometry, othewise None
-        super(VectorData, self).__init__(out, deep_copy, self.geometry.dimension_labels, geometry = self.geometry)
-
-
-class VectorGeometry(object):
-    '''Geometry describing VectorData to contain 1D array'''
-    RANDOM = 'random'
-    RANDOM_INT = 'random_int'
-
-    @property
-    def dtype(self):
-        return self._dtype
-
-    @dtype.setter
-    def dtype(self, val):
-        self._dtype = val      
-        
-    def __init__(self, 
-                 length, **kwargs):
-        
-        self.length = int(length)
-        self.shape = (length, )
-        self.dtype = kwargs.get('dtype', numpy.float32)
-        self.dimension_labels = kwargs.get('dimension_labels', None)
-        
-    def clone(self):
-        '''returns a copy of VectorGeometry'''
-        return copy.deepcopy(self)
-
-    def copy(self):
-        '''alias of clone'''
-        return self.clone()
-
-    def __eq__(self, other):
-
-        if not isinstance(other, self.__class__):
-            return False
-        
-        if self.length == other.length \
-            and self.shape == other.shape \
-            and self.dimension_labels == other.dimension_labels:
-            return True
-        return False
-
-    def __str__ (self):
-        repres = ""
-        repres += "Length: {0}\n".format(self.length)
-        repres += "Shape: {0}\n".format(self.shape)
-        repres += "Dimension_labels: {0}\n".format(self.dimension_labels)
-
-        return repres
-
-    def allocate(self, value=0, **kwargs):
-        '''allocates an VectorData according to the size expressed in the instance
-        
-        :param value: accepts numbers to allocate an uniform array, or a string as 'random' or 'random_int' to create a random array or None.
-        :type value: number or string, default None allocates empty memory block
-        :param dtype: numerical type to allocate
-        :type dtype: numpy type, default numpy.float32
-        :param seed: seed for the random number generator
-        :type seed: int, default None
-        :param max_value: max value of the random int array
-        :type max_value: int, default 100'''
-
-        dtype = kwargs.get('dtype', self.dtype)
-        # self.dtype = kwargs.get('dtype', numpy.float32)
-        out = VectorData(geometry=self.copy(), dtype=dtype)
-        if isinstance(value, Number):
-            if value != 0:
-                out += value
-        else:
-            if value == VectorGeometry.RANDOM:
-                seed = kwargs.get('seed', None)
-                if seed is not None:
-                    numpy.random.seed(seed) 
-                out.fill(numpy.random.random_sample(self.shape))
-            elif value == VectorGeometry.RANDOM_INT:
-                seed = kwargs.get('seed', None)
-                if seed is not None:
-                    numpy.random.seed(seed)
-                max_value = kwargs.get('max_value', 100)
-                r = numpy.random.randint(max_value,size=self.shape, dtype=numpy.int32)
-                out.fill(numpy.asarray(r, dtype=self.dtype))             
-            elif value is None:
-                pass
-            else:
-                raise ValueError('Value {} unknown'.format(value))
-        return out
-
 class DataOrder():
 
     ENGINES = ['astra','tigre','cil']
